gallery: undo complicated placeholder for simpler solution
